+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.28.0 (??)
 - Feature: new cosmetic finder to fix cursive within links
 - Save more details about a reviewed replacement. Save if the replacement has been modified or not during the review, along with the timestamp and the revisions before and after in case we need to revert it.
+- Ordinal finder: suggest always the feminine fix
 - Fix: save the timestamp returned by Wikipedia API instead of the local date
 - Fix: some finders stop finding after the first false positive
 - Fix: ignore ordinals preceded by letter